+++ packages/medusa-source-shopify/src/services/__mocks__/shopify-redis.js
@@ -1,0 +1,8 @@
+export const ShopifyRedisServiceMock = {
+  addIgnore: jest.fn().mockImplementation((_id, _event) => {
+    return Promise.resolve()
+  }),
+  shouldIgnore: jest.fn().mockImplementation((_id, _event) => {
+    return false
+  }),
+}