+++ packages/medusa-source-shopify/src/utils/create-client.js
@@ -1,0 +1,7 @@
+import Shopify from "@shopify/shopify-api"
+
+export const createClient = (options) => {
+  const { domain, password } = options
+
+  return new Shopify.Clients.Rest(`${domain}.myshopify.com`, password)
+}