+++ packages/medusa/src/api/routes/admin/shipping-options/index.ts
@@ -1,4 +1,6 @@
 import { Router } from "express"
+import { ShippingOption } from "../../../.."
+import { DeleteResponse } from "../../../../types/common"
 import middlewares from "../../../middlewares"
 
 const route = Router()
@@ -43,3 +45,19 @@ export const defaultFields = [
 ]
 
 export const defaultRelations = ["region", "profile", "requirements"]
+
+export type AdminShippingOptionsListRes = {
+  shipping_options: ShippingOption[]
+}
+
+export type AdminShippingOptionsRes = {
+  shipping_option: ShippingOption
+}
+
+export type AdminShippingOptionsDeleteRes = DeleteResponse
+
+export * from "./create-shipping-option"
+export * from "./delete-shipping-option"
+export * from "./get-shipping-option"
+export * from "./list-shipping-options"
+export * from "./update-shipping-option"