+++ packages/medusa/src/api/routes/store/carts/refresh-payment-session.ts
@@ -1,7 +1,7 @@
 import { CartService } from "../../../../services"
 
 /**
- * @oas [post] /carts/{id}/payment-sessions/{provider_id}
+ * @oas [post] /carts/{id}/payment-sessions/{provider_id}/refresh
  * operationId: PostCartsCartPaymentSessionsSession
  * summary: Refresh a Payment Session
  * description: "Refreshes a Payment Session to ensure that it is in sync with the Cart - this is usually not necessary."